Refactoring after V1
Went back in to clean up the code before adding/changing features. Huge
amount of redundancy removed now that I'm using classes and extensions
correctly. Program was not even adding attribute to cards, let alone
comparing it, so I added both of those. Added variables to the Screen
class for scaling components on them, thought is to use it to make
Component creation a lot easier/more readable. Started the transfer from
hardcoded component creation to files housing all of the information,
again to make creation and modification way easier, and to open up the
possibility for user intervention in the future as well.