Multiple Random Decks Working
You are now able to generate multiple decks at once, all following the
settings that you chose. Buttons created when you make the random decks
now also work to switch between the decks, and decks are able to span
multiple pages and you can still navigate through each one. Also removed
the ydk files from the tracking.

Also found an issue with the addCards fuction of Deck, it was return
true when it had as many/more cards than the user said a deck should
have, and the while loop that tries to generate deck(s) was assuming
addCards returned true if more cards needed to be added. Fixed that in
the addCards function, which sped it up quite a bit. ImageGridCollection
appears to be memory safe (can jump back and forth between settings and
decks without it keeping the previous images), which is great!