Library updating
SimpleGUI has been updated to now allow for dynamic x, y, w, h values,
which broke some components of the randomizer. All have been updated to
work with the library. ImageGrid and Collection seem to be slower now,
which would make sense since they are getting a lot of instances of
Screen passed to them (with every Card added). Will have to look into
this.