Library updated to reduce redundancy
Library was updated so that instead of passing a copy of Screen every
draw/mouseOver function call, it is set as a variable once the Component
is added to a Screen. This cut down on unneccesary copies of Screen
being sent all over the place, and increased performance. Image Grid
(Collection) had to be updated to fit this new change.